refactored code to use github-api-pages plugin
