+++ packages/documentation/gatsby-config.js
@@ -57,5 +57,6 @@ module.exports = {
     },
     `gatsby-transformer-react-docgen`,
     `gatsby-transformer-remark`,
+    `github-api-pages`,
   ],
 }
