+++ packages/documentation/plugins/github-api-pages/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "github-api-pages",
+  "version": "1.0.0",
+  "description": "",
+  "main": "gatsby-node.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
